+++ Changelog.md
@@ -11,6 +11,7 @@
 * In codebooks, renaming some variables and attributes regarding the use of kmeans initialization, for clarity.
 * Documentation and clarification of 'batched_bincount' function.
 * Change the level at which the check on whether the embeddings have initialized or not, that is now before calling the method initialize_embeddings and not inside it. Improve performance and clarity.
+* Adapt change of name for attribute embed to embeddings in ResidualVQ.
 
 ## [Version 1.16.0]